Added tag jdk-20+17 for changeset 79ccc791

+++ src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/debugger/windbg/amd64/WindbgAMD64Thread.java
@@ -36,7 +36,7 @@ class WindbgAMD64Thread implements ThreadProxy {
   // The address argument must be the address of the OSThread::_thread_id
   WindbgAMD64Thread(WindbgDebugger debugger, Address addr) {
     this.debugger = debugger;
-    this.sysId    = (long)addr.getCIntegerAt(0, 4, true);
+    this.sysId    = addr.getCIntegerAt(0, 4, true);
     gotID         = false;
   }
